Add database snapshotting to juju-restore
Before restoring the dump in the backup we snapshot the current
database by stopping mongo, copying the /var/lib/juju/db directory
on all controller machines and restarting it.

If the restore is successful, these snapshots are discarded on each
machine. If it fails for some reason, the snapshots are restored by
copying the snapshots back to /var/lib/juju/db (after stopping the
database).

This isn't complete yet. At the moment we don't take the step of restarting the
jujud agents yet after restoring the snapshots in the case of a failed
restore. If the restore is also a version downgrade that fails to run
the agent update script on some of the nodes, we don't revert the
upgrade script on the nodes where it succeeded.

This also needs unit tests for the Snapshotter and the points where
it's used.